fix UI add login, add note, split file
