+++ popup-src/event-handlers.js
@@ -3,7 +3,7 @@ import { renderBySelection } from './UI/vault-manager.js';
 import {
   renderAddNoteUI,
   renderAddLoginUI,
-} from './UI/lock/ui-renderer.js';
+} from './ui-renderer.js';
 import { renderChangePasswordUI } from './settings.js';
 import { handleGeneratePassword } from './password-generator.js';
 import {
